+++ src/app/index.js
@@ -1,10 +1,13 @@
 import { createApp } from './app';
 import Vitessce from './Vitessce';
 import { encodeConfInUrl, decodeURLParamsToConf } from './export-utils';
+import { registerPluginFileType, registerPluginViewType } from './plugins';
 
 export {
   createApp,
   Vitessce,
   encodeConfInUrl,
   decodeURLParamsToConf,
+  registerPluginFileType,
+  registerPluginViewType,
 };