+++ src/index.js
@@ -7,6 +7,8 @@ import {
   Vitessce,
   encodeConfInUrl,
   decodeURLParamsToConf,
+  registerPluginFileType,
+  registerPluginViewType,
 } from './app';
 import {
   VitessceConfig,
@@ -19,6 +21,26 @@ import {
   FileType,
   CoordinationType,
 } from './app/constants';
+// For plugin views to use:
+import TitleInfo from './components/TitleInfo';
+import { useReady, useUrls } from './components/hooks';
+import {
+  useDescription,
+  useCellsData,
+  useCellSetsData,
+  useExpressionMatrixData,
+  useGeneSelection,
+  useExpressionAttrs,
+  useMoleculesData,
+  useNeighborhoodsData,
+  useRasterData,
+  useGenomicProfilesData,
+} from './components/data-hooks';
+import {
+  useCoordination,
+  useLoaders,
+} from './app/state/hooks';
+
 
 export {
   Heatmap,
@@ -36,4 +58,24 @@ export {
   DataType,
   FileType,
   CoordinationType,
+  // Plugin registration functions
+  registerPluginFileType,
+  registerPluginViewType,
+  // Exports for plugins
+  // (not guaranteed to be compatible across different Vitessce versions)
+  TitleInfo,
+  useReady,
+  useUrls,
+  useCoordination,
+  useLoaders,
+  useDescription,
+  useCellsData,
+  useCellSetsData,
+  useExpressionMatrixData,
+  useGeneSelection,
+  useExpressionAttrs,
+  useMoleculesData,
+  useNeighborhoodsData,
+  useRasterData,
+  useGenomicProfilesData,
 };